+++ Meraki.Api/MerakiClient_SerialNumberInfo.cs
@@ -177,6 +177,7 @@ public partial class MerakiClient
 		{ "Q3LR", "MS130-24P" },
 		{ "Q3LS", "MS130-24X" },
 		{ "Q3LV", "MS130-48X" },
+		{ "Q3LW", "MS130-8P-I" },
 
 		{ "Q4AA", "MS120-8" },
 		{ "Q4AB", "MS120-8LP" },